CAPTION GENERATOR/requirements.txt
openai-whisper>=20231117
moviepy>=1.0.3
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.66.0
deep-translator>=1.11.4
langdetect>=1.0.9
torch>=2.0.0
torchaudio>=2.0.0


